+++ src/permissions.ts
@@ -1,8 +1,3 @@
-import {
-  findMask,
-  normalizeMask,
-  type PermissionMask,
-} from "./permission-mask.js";
 import { createPermissionStatus } from "./permission-status.js";
 import type {
   PermissionAccessStatus,
@@ -32,39 +27,6 @@ export type _DocsTypes =
  *   {@link globalThis.Permissions | Permissions} API.
  */
 export interface PermissionsParameters {
-  /**
-   * Permission masks to apply when mapping {@link PermissionAccessStatus}
-   * values to {@link PermissionState} values for various permissions.
-   *
-   * Internally, this library stores more granular statuses for permissions than
-   * what is exposed via the W3C {@link globalThis.Permissions | Permissions}
-   * API. This mapping defines how the internal {@link PermissionAccessStatus}
-   * values are exposed via the fake
-   * {@link globalThis.Permissions | Permissions} API's {@link PermissionState}
-   * values.
-   *
-   * For example, some browsers support allowing one-time access to a
-   * permission. This usually results in the {@link PermissionState} remaining
-   * as {@link PermissionState | `"prompt"`} even after access has been allowed
-   * or denied. In order to emulate this behavior, the default behavior of this
-   * library is to use a mask that maps both
-   * {@link PermissionAccessStatusAllowed | `"ALLOWED"`} and
-   * {@link PermissionAccessStatusDenied | `"DENIED"`} to
-   * {@link PermissionState | `"prompt"`}.
-   *
-   * If no explicit mappings are provided, the default mapping is:
-   *
-   * | {@link PermissionAccessStatus}                                                 | {@link PermissionState}               |
-   * | :----------------------------------------------------------------------------- | :------------------------------------ |
-   * | {@link PermissionAccessStatusPrompt | `"PROMPT"`}                              | {@link PermissionState | `"prompt"`}  |
-   * | {@link PermissionAccessStatusGranted | `"GRANTED"`}                            | {@link PermissionState | `"granted"`} |
-   * | {@link PermissionAccessStatusBlocked | `"BLOCKED"`}                            | {@link PermissionState | `"denied"`}  |
-   * | {@link PermissionAccessStatusBlockedAutomatically | `"BLOCKED_AUTOMATICALLY"`} | {@link PermissionState | `"denied"`}  |
-   * | {@link PermissionAccessStatusAllowed | `"ALLOWED"`}                            | {@link PermissionState | `"prompt"`}  |
-   * | {@link PermissionAccessStatusDenied | `"DENIED"`}                              | {@link PermissionState | `"prompt"`}  |
-   */
-  masks?: Map<PermissionDescriptor, Partial<PermissionMask>>;
-
   /**
    * The permission store to use.
    */
@@ -96,11 +58,10 @@ export class Permissions {
    * @internal
    * @deprecated Use {@link createPermissions} instead.
    */
-  constructor({ masks = new Map(), permissionStore }: PermissionsParameters) {
+  constructor({ permissionStore }: PermissionsParameters) {
     if (!canConstruct) throw new TypeError("Illegal constructor");
     canConstruct = false;
 
-    this.#masks = masks;
     this.#permissionStore = permissionStore;
   }
 
@@ -131,15 +92,11 @@ export class Permissions {
 
     return createPermissionStatus({
       descriptor,
-      mask: normalizeMask(
-        findMask(this.#permissionStore, this.#masks, descriptor) ?? {},
-      ),
       permissionStore: this.#permissionStore,
     });
   }
 
   readonly [Symbol.toStringTag] = "Permissions";
 
-  readonly #masks: Map<PermissionDescriptor, Partial<PermissionMask>>;
   readonly #permissionStore: PermissionStore;
 }